+++ FlintCore/Constraints/Permissions/LocationUsage.swift
@@ -1,0 +1,14 @@
+//
+//  LocationUsage.swift
+//  FlintCore
+//
+//  Created by Marc Palmer on 01/05/2018.
+//  Copyright © 2018 Montana Floss Co. Ltd. All rights reserved.
+//
+
+import Foundation
+
+public enum LocationUsage {
+    case always
+    case whenInUse
+}
