+++ FlintCore/Constraints/Permissions/Permission.swift
@@ -1,0 +1,12 @@
+//
+//  Permissions.swift
+//  FlintCore
+//
+//  Created by Marc Palmer on 28/04/2018.
+//  Copyright © 2018 Montana Floss Co. Ltd. All rights reserved.
+//
+
+import Foundation
+
+public protocol Permission: Hashable {
+}
