+++ FlintCore/Constraints/Permissions/PermissionStatus.swift
@@ -1,0 +1,17 @@
+//
+//  PermissionStatus.swift
+//  FlintCore
+//
+//  Created by Marc Palmer on 01/05/2018.
+//  Copyright © 2018 Montana Floss Co. Ltd. All rights reserved.
+//
+
+import Foundation
+
+public enum PermissionStatus {
+    case unknown
+    case authorized
+    case denied
+    case restricted
+    case unsupported
+}
